2 more utils
